feat(utils): refactor
